move functions into classify
